gpu_momentum_trading/requirements.txt
# GPU Momentum Trading System Requirements

# GPU Computing
cupy-cuda12x
numba

# Trading APIs
alpaca-py

# Async and Networking
aiohttp
websockets

# Data Processing
numpy
pytz

# Utilities
python-dotenv